+++ lib/protoparser/graphite/parser_test.go
@@ -288,6 +288,23 @@ func TestRowsUnmarshalSuccess(t *testing.T) {
 		}},
 	})
 
+	// Whitespace after the timestamp
+	// See https://github.com/VictoriaMetrics/VictoriaMetrics/issues/1865
+	f("foo.baz 125 1789 \na 1.34 567\t  ", &Rows{
+		Rows: []Row{
+			{
+				Metric:    "foo.baz",
+				Value:     125,
+				Timestamp: 1789,
+			},
+			{
+				Metric:    "a",
+				Value:     1.34,
+				Timestamp: 567,
+			},
+		},
+	})
+
 }
 
 func Test_streamContext_Read(t *testing.T) {